bpf: Support encapsulation for packets received from a netdev
This extends bpf_netdev to perform encapsulation if the destination is
within the cluster prefix but not a local endpoint. This allows for a
local process to reach endpoints which reside on other nodes.

If the packet is being received from the wire, the identity is inherited
as normal. If the packet is coming from the host, the skb->mark can be
used to define the source identity of the packet.

It also compiles bpf_netdev.c to bpf_host.o for use on the cilium_host
interface so we can pass in a special define FROM_HOST which indicates
that bpf_netdev is being run in the context for packets from the host.

Signed-off-by: Thomas Graf <thomas@cilium.io>